+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Change `RegistrationPhoneScreen` to use view effects
 - Change `ScanSimpleIdScreen` to use Mobius view effects
 - Handle exception when syncing reports in `ConfirmFacilityChangeEffectHandler`
+- Show appointment results in contact patient sheet for patient with no phone number
 
 ### Changes
 - Ask users for their name and job designation on sign up